engine: extract esm/folder_up.js
