+++ engine/esm/index.js
@@ -212,6 +212,7 @@ export { Lineset, Linepoint, Constellations } from "./constellations.js";
 export { SpaceTimeController } from "./space_time_controller.js";
 export { KeplerianElements, BodyAngles, Planets } from "./planets.js";
 export { Place } from "./place.js";
+export { FolderUp } from "./folder_up.js";
 
 export {
     PointType,
@@ -248,7 +249,6 @@ export {
     Pointing,
     Folder,
     FolderBrowser,
-    FolderUp,
     Grids,
     ViewMoverKenBurnsStyle,
     KeplerVertex,
