+++ engine/esm/transpiled.js
@@ -144,6 +144,7 @@ import { Text3d, Text3dBatch } from "./sky_text.js";
 import { SpaceTimeController } from "./space_time_controller.js";
 import { Planets } from "./planets.js";
 import { Place } from "./place.js";
+import { FolderUp } from "./folder_up.js";
 
 
 // wwtlib.PointType
@@ -1701,75 +1702,6 @@ var FolderBrowser$ = {
 };
 
 registerType("FolderBrowser", [FolderBrowser, FolderBrowser$, null]);
-
-// wwtlib.FolderUp
-
-export function FolderUp() {
-  this.parent = null;
-  this._bounds = new Rectangle();
-}
-
-var FolderUp$ = {
-  get_name: function () {
-    return 'Up Level';
-  },
-
-  get_thumbnail: function () {
-    return this._thumbnail;
-  },
-
-  set_thumbnail: function (value) {
-    this._thumbnail = value;
-    return value;
-  },
-
-  get_thumbnailUrl: function () {
-    return URLHelpers.singleton.engineAssetUrl('thumb_folderup.jpg');
-  },
-
-  set_thumbnailUrl: function (value) {
-    return value;
-  },
-
-  get_bounds: function () {
-    return this._bounds;
-  },
-
-  set_bounds: function (value) {
-    this._bounds = value;
-    return value;
-  },
-
-  get_isImage: function () {
-    return false;
-  },
-
-  get_isTour: function () {
-    return false;
-  },
-
-  get_isFolder: function () {
-    return false;
-  },
-
-  get_isCloudCommunityItem: function () {
-    return false;
-  },
-
-  get_readOnly: function () {
-    return false;
-  },
-
-  get_children: function () {
-    if (this.parent == null) {
-      return [];
-    } else {
-      return this.parent.get_children();
-    }
-  }
-};
-
-registerType("FolderUp", [FolderUp, FolderUp$, null, IThumbnail]);
 
 // wwtlib.Grids
